requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.18
sqlalchemy==2.0.23
boto3==1.34.0
mangum==0.17.0
python-dotenv==1.0.0